Split s3 testutils into new module
Refactor some s3 fixture functions so they're in the fixture
